move all to new table of products block
