--- blocks/all-products/all-products.css
@@ -1,27 +1,0 @@
-.all-products {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: space-between;
-}
-
-.all-products > div {
-    width: calc(100% * (1/2) );
-    display: flex;
-    flex-direction: column;
-    text-align: center;
-    margin-bottom: 30px;
-}
-
-@media (min-width: 1040px) {
-    .all-products > div {
-        width: calc(100% * (1 / 4));
-    }
-}
-
-.all-products a, .all-products p, .all-products h3, .all-products a.button {
-    margin: 0;
-}
-
-.all-products-container {
-    background: url('/media/apples-top-1024x220-1.png?width=2000&format=webply&optimize=medium') no-repeat;
-}
