+++ styles/theme-spirits.css
@@ -1,23 +1,3 @@
-.spirits {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: space-between;
-}
-
-.spirits > div {
-    width: calc(100% * (1/2) );
-    display: flex;
-    flex-direction: column;
-    text-align: center;
-    margin-bottom: 30px;
-}
-
-@media (min-width: 1040px) {
-    .spirits > div {
-        width: calc(100% * (1 / 4));
-    }
-}
-
 /* the yellow title under each picture */
 .spirits p:nth-of-type(2) {
     color: var(--header-background-color);
@@ -53,14 +33,14 @@
 }
 
 main {
-    background: url('/media/spirits-background-left.png');
+    background: url('/media/spirits-background-left.png?width=2000&format=webply&optimize=medium');
     background-position: top left;
     background-repeat: no-repeat;
     background-color: transparent !important;
 }
 
-.spirits-container {
-    background: url('/media/spirits-background-right.png');
+.section {
+    background: url('/media/spirits-background-right.png?width=2000&format=webply&optimize=medium');
     background-position: bottom right;
     background-attachment: fixed;
     background-repeat: no-repeat;
